Fix issues when size_t != unsigned long
This is a patch due to Benjamin Burton. To quote him:

This fixes several compile errors in the GMPXX-related code on systems where
size_t != unsigned long. This happens in particular on 64-bit windows. (You
might not be seeing this because you use MPIR on windows, but regina uses GMP
everywhere so probably I am building this code on additional platforms where
you are not.)

The compile error is related to overloaded functions in GMP that C++ cannot
choose between. I believe the problem is that you are passing size_t arguments
to GMP functions, but these GMP functions do not take size_t arguments. On
normal platforms this is not a problem since size_t == unsigned long, and GMP
functions *do* take unsigned long arguments. However, if size_t != unsigned
long then the compiler needs to cast size_t to a different type that the GMP
functions do accept, and since GMP offers several types to choose from, the
compiler fails because it needs to be told which typecast to choose.

In my patches, I explicitly cast to an unsigned long, which is the largest
native integer type that GMP will accept. This is a cast down (size_t is 8
bytes but long is 4); however, I am (naively?) assuming that these arguments
are indeed sizes of structures (e.g., matrix ranks, vector lengths,
dimensions, row indices, etc.) and so if they ever exceed MAX_LONG then there
are much bigger problems to deal with.

+++ source/libnormaliz/full_cone.cpp
@@ -3075,7 +3075,7 @@ void Full_Cone<Integer>::find_bottom_facets() {
     INTERRUPT_COMPUTATION_BY_EXCEPTION
 
     Matrix<Integer> BottomFacets(0, dim);
-    vector<Integer> BottomDegs(0, dim);
+    vector<Integer> BottomDegs(0, static_cast<unsigned long>(dim));
     if (!isComputed(ConeProperty::SupportHyperplanes)) {
         Support_Hyperplanes = Matrix<Integer>(0, dim);
         nrSupport_Hyperplanes = 0;
@@ -6106,7 +6106,7 @@ void Full_Cone<Integer>::compute_class_group() {  // from the support hyperplane
     Matrix<Integer> Trans = Support_Hyperplanes;  // .transpose();
     size_t rk;
     Trans.SmithNormalForm(rk);
-    ClassGroup.push_back(Support_Hyperplanes.nr_of_rows() - rk);
+    ClassGroup.push_back(static_cast<unsigned long>(Support_Hyperplanes.nr_of_rows() - rk));
     for (size_t i = 0; i < rk; ++i)
         if (Trans[i][i] != 1)
             ClassGroup.push_back(Trans[i][i]);
